requirements.txt
asgiref==3.10.0
brotli==1.2.0
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
filetype==1.2.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.38.0
whitenoise==6.11.0
django-storages==1.14.2



